split lakectl cmds - part 1
