+++ cmd/lakectl/cmd/common_helpers.go
@@ -2,7 +2,6 @@ package cmd
 
 import (
 	"bytes"
-	"encoding/base64"
 	"encoding/json"
 	"errors"
 	"fmt"
@@ -10,7 +9,6 @@ import (
 	"net/http"
 	"os"
 	"reflect"
-	"regexp"
 	"strconv"
 	"strings"
 	"text/template"
@@ -18,22 +16,15 @@ import (
 
 	"github.com/jedib0t/go-pretty/v6/table"
 	"github.com/jedib0t/go-pretty/v6/text"
+	"github.com/manifoldco/promptui"
+	"github.com/spf13/pflag"
 	"github.com/treeverse/lakefs/pkg/api"
 	"github.com/treeverse/lakefs/pkg/api/helpers"
 	"github.com/treeverse/lakefs/pkg/uri"
 	"golang.org/x/term"
 )
 
-var (
-	isTerminal       = true
-	noColorRequested = false
-	verboseMode      = false
-)
-
-// ErrInvalidValueInList is an error returned when a parameter of type list contains an empty string
-var ErrInvalidValueInList = errors.New("empty string in list")
-
-var accessKeyRegexp = regexp.MustCompile(`^AKIA[I|J][A-Z0-9]{14}Q$`)
+var isTerminal = true
 
 const (
 	PathDelimiter = "/"
@@ -57,7 +48,7 @@ const resourceListTemplate = `{{.Table | table -}}
 //nolint:gochecknoinits
 func init() {
 	// disable colors if we're not attached to interactive TTY.
-	// when environment variable is set we use it to control interactive mode
+	// when an environment variable is set, we use it to control interactive mode
 	// otherwise we will try to detect based on the standard output
 	interactiveVal := os.Getenv(LakectlInteractive)
 	if interactiveVal != "" {
@@ -182,13 +173,14 @@ func WriteIfVerbose(tpl string, data interface{}) {
 	}
 }
 
-func Die(err string, code int) {
-	WriteTo(DeathMessage, struct{ Error string }{err}, os.Stderr)
+func Die(errMsg string, code int) {
+	WriteTo(DeathMessage, struct{ Error string }{Error: errMsg}, os.Stderr)
 	os.Exit(code)
 }
 
 func DieFmt(msg string, args ...interface{}) {
-	Die(fmt.Sprintf(msg, args...), 1)
+	errMsg := fmt.Sprintf(msg, args...)
+	Die(errMsg, 1)
 }
 
 type APIError interface {
@@ -213,10 +205,6 @@ func DieErr(err error) {
 		WriteTo(DeathMessage, ErrData{Error: err.Error()}, os.Stderr)
 	}
 	os.Exit(1)
-}
-
-type StatusCoder interface {
-	StatusCode() int
 }
 
 func RetrieveError(response interface{}, err error) error {
@@ -260,10 +248,6 @@ func DieOnHTTPError(httpResponse *http.Response) {
 	if err != nil {
 		DieErr(err)
 	}
-}
-
-func Fmt(msg string, args ...interface{}) {
-	fmt.Printf(msg, args...)
 }
 
 func PrintTable(rows [][]interface{}, headers []interface{}, paginator *api.Pagination, amount int) {
@@ -331,15 +315,99 @@ func MustParsePathURI(name, s string) *uri.URI {
 	return u
 }
 
-func IsValidAccessKeyID(accessKeyID string) bool {
-	return accessKeyRegexp.MatchString(accessKeyID)
+const (
+	AutoConfirmFlagName     = "yes"
+	AutoConfigFlagShortName = "y"
+	AutoConfirmFlagHelp     = "Automatically say yes to all confirmations"
+
+	StdinFileName = "-"
+)
+
+func AssignAutoConfirmFlag(flags *pflag.FlagSet) {
+	flags.BoolP(AutoConfirmFlagName, AutoConfigFlagShortName, false, AutoConfirmFlagHelp)
 }
 
-func IsValidSecretAccessKey(secretAccessKey string) bool {
-	return IsBase64(secretAccessKey) && len(secretAccessKey) == 40
+func Confirm(flags *pflag.FlagSet, question string) (bool, error) {
+	yes, err := flags.GetBool(AutoConfirmFlagName)
+	if err == nil && yes {
+		// got auto confirm flag
+		return true, nil
+	}
+	prm := promptui.Prompt{
+		Label:     question,
+		IsConfirm: true,
+	}
+	_, err = prm.Run()
+	if err != nil {
+		return false, err
+	}
+	return true, nil
 }
 
-func IsBase64(s string) bool {
-	_, err := base64.StdEncoding.DecodeString(s)
-	return err == nil
+// nopCloser wraps a ReadSeekCloser to ignore calls to Close().  It is io.NopCloser (or
+// ioutils.NopCloser) for Seeks.
+type nopCloser struct {
+	io.ReadSeekCloser
+}
+
+func (nc *nopCloser) Close() error {
+	return nil
+}
+
+// deleteOnClose wraps a File to be a ReadSeekCloser that deletes itself when closed.
+type deleteOnClose struct {
+	*os.File
+}
+
+func (d *deleteOnClose) Close() error {
+	if err := os.Remove(d.Name()); err != nil {
+		_ = d.File.Close() // "Only" file descriptor leak if close fails (but data might stay).
+		return fmt.Errorf("delete on close: %w", err)
+	}
+	return d.File.Close()
+}
+
+// OpenByPath returns a reader from the given path.
+// If the path is "-", it consumes Stdin and
+// opens a readable copy that is either deleted (POSIX) or will delete itself on close
+// (non-POSIX, notably WINs).
+func OpenByPath(path string) (io.ReadSeekCloser, error) {
+	if path != StdinFileName {
+		return os.Open(path)
+	}
+
+	// check if stdin is seekable
+	_, err := os.Stdin.Seek(0, io.SeekCurrent)
+	if err == nil {
+		return &nopCloser{ReadSeekCloser: os.Stdin}, nil
+	}
+
+	temp, err := os.CreateTemp("", "lakectl-stdin")
+	if err != nil {
+		return nil, fmt.Errorf("create temporary file to buffer stdin: %w", err)
+	}
+	if _, err = io.Copy(temp, os.Stdin); err != nil {
+		return nil, fmt.Errorf("copy stdin to temporary file: %w", err)
+	}
+	if _, err = temp.Seek(0, io.SeekStart); err != nil {
+		return nil, fmt.Errorf("rewind temporary copied file: %w", err)
+	}
+	// Try to delete the file.  This will fail on Windows, we shall try to
+	// delete on close anyway.
+	if os.Remove(temp.Name()) != nil {
+		return &deleteOnClose{File: temp}, nil
+	}
+	return temp, nil
+}
+
+// Must return the call value or die with err if err is not nil
+func Must[T any](v T, err error) T {
+	if err != nil {
+		DieErr(err)
+	}
+	return v
+}
+
+func Fmt(format string, a ...interface{}) string {
+	return fmt.Sprintf(format, a...)
 }
