+++ cmd/lakectl/cmd/merge.go
@@ -36,7 +36,7 @@ var mergeCmd = &cobra.Command{
 		client := getClient()
 		sourceRef := MustParseRefURI("source ref", args[0])
 		destinationRef := MustParseRefURI("destination ref", args[1])
-		strategy := MustString(cmd.Flags().GetString("strategy"))
+		strategy := Must(cmd.Flags().GetString("strategy"))
 		Fmt("Source: %s\nDestination: %s\n", sourceRef.String(), destinationRef)
 		if destinationRef.Repository != sourceRef.Repository {
 			Die("both references must belong to the same repository", 1)
